get_journalmetrics: Add journal rankings from scimago
This could be useful for having more info about each publication (ie,
journal stats) to be able to further investigate scientific impact. The
function get_journalmetrics is almost the same as get_impactfactor (with
simply the data changed). I am not sure how to include the scimago
dataset, so I based it on what was done with the impact factor dataset.

Two lines of enhancement include:
1. Merge get_impactfactor into the get_journalmetrics function
2. Enhance the way it matches journal names between the dataset and the
scholar profile. The current `agrep` methods generates a non-negligible
amount of errors.